chore: move integration connection interface
